\name{vp-package}
\alias{vp-package}
\alias{vp}
\docType{package}
\title{
\packageTitle{vp}
}
\description{
\packageDescription{vp}
}
\details{

The \code{poker.pval()} function maps p values to video poker hands.

Use \code{toggle_stars_poker()} to replace the built-in p-value formater with poker (and disable stars).

}
\author{
\packageAuthor{vp}

Maintainer: \packageMaintainer{vp}
}
\keyword{ package }
\seealso{
  \code{\link{format.pval}}  if base R
}
\examples{
\dontrun{

}
}
